% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_appname_from_url.R
\name{get_appname_from_url}
\alias{get_appname_from_url}
\title{Read path to appname on rsconnect}
\usage{
get_appname_from_url(session = getDefaultReactiveDomain())
}
\description{
Read path to appname on rsconnect
}
