% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msalUserManagementModule.R
\name{msalUserManagementModule}
\alias{msalUserManagementModule}
\title{MSAL User Management Module}
\usage{
msalUserManagementModule(input, output, session, .user, appname)
}
\arguments{
\item{input}{input modalities of the module}

\item{output}{output modalities of the module}

\item{session}{R session object of the module}

\item{.user}{user object, connection to the shintousers database and functions necessary}

\item{appname}{appname for which the user can see connections}
}
\description{
A light version from our internal shintousers-app, so admin users can edit rights within the app
}
