% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msalUserManagementModule.R
\name{msalUserManagementUI}
\alias{msalUserManagementUI}
\title{MSAL User Management UI}
\usage{
msalUserManagementUI(id)
}
\arguments{
\item{id}{ID of the module}
}
\description{
A light version from our internal shintousers-app, so admin users can edit rights within the app
}
