% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msal_users_r6_object.R
\name{shintoMSALUser}
\alias{shintoMSALUser}
\title{A User object for use in Shinto apps}
\value{
Returns (invisibly) the last login information
}
\description{
Make a database connection to 'shintousers' and access all kinds of
methods for user management. In Shiny apps, use `get_and_set_last_login`, for example.
}
\section{Super class}{
\code{shintodb::NA} -> \code{ShintoMSALUser}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{The DB connection to `shintousers`}

\item{\code{schema}}{The postgres schema in the DB}

\item{\code{dbusername}}{The DB user name}

\item{\code{userid}}{The application (shiny) user name}

\item{\code{appname}}{The  application name}

\item{\code{appversion}}{Version of the app (optionally read from VERSION)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShintoMSALUser-new}{\code{shintoMSALUser$new()}}
\item \href{#method-ShintoMSALUser-to_json}{\code{shintoMSALUser$to_json()}}
\item \href{#method-ShintoMSALUser-from_json}{\code{shintoMSALUser$from_json()}}
\item \href{#method-ShintoMSALUser-add_msal_user}{\code{shintoMSALUser$add_msal_user()}}
\item \href{#method-ShintoMSALUser-get_name}{\code{shintoMSALUser$get_name()}}
\item \href{#method-ShintoMSALUser-set_user_name}{\code{shintoMSALUser$set_user_name()}}
\item \href{#method-ShintoMSALUser-get_email}{\code{shintoMSALUser$get_email()}}
\item \href{#method-ShintoMSALUser-set_user_email}{\code{shintoMSALUser$set_user_email()}}
\item \href{#method-ShintoMSALUser-get_comment}{\code{shintoMSALUser$get_comment()}}
\item \href{#method-ShintoMSALUser-set_comment}{\code{shintoMSALUser$set_comment()}}
\item \href{#method-ShintoMSALUser-get_last_login}{\code{shintoMSALUser$get_last_login()}}
\item \href{#method-ShintoMSALUser-set_last_login}{\code{shintoMSALUser$set_last_login()}}
\item \href{#method-ShintoMSALUser-get_and_set_last_login}{\code{shintoMSALUser$get_and_set_last_login()}}
\item \href{#method-ShintoMSALUser-app_has_user}{\code{shintoMSALUser$app_has_user()}}
\item \href{#method-ShintoMSALUser-list_application_users}{\code{shintoMSALUser$list_application_users()}}
\item \href{#method-ShintoMSALUser-set_user_attributes}{\code{shintoMSALUser$set_user_attributes()}}
\item \href{#method-ShintoMSALUser-get_user_attributes}{\code{shintoMSALUser$get_user_attributes()}}
\item \href{#method-ShintoMSALUser-get_role}{\code{shintoMSALUser$get_role()}}
\item \href{#method-ShintoMSALUser-set_role}{\code{shintoMSALUser$set_role()}}
\item \href{#method-ShintoMSALUser-get_user_active_inactive}{\code{shintoMSALUser$get_user_active_inactive()}}
\item \href{#method-ShintoMSALUser-set_user_active_inactive}{\code{shintoMSALUser$set_user_active_inactive()}}
\item \href{#method-ShintoMSALUser-disable_user}{\code{shintoMSALUser$disable_user()}}
\item \href{#method-ShintoMSALUser-enable_user}{\code{shintoMSALUser$enable_user()}}
\item \href{#method-ShintoMSALUser-get_group}{\code{shintoMSALUser$get_group()}}
\item \href{#method-ShintoMSALUser-is_in_group}{\code{shintoMSALUser$is_in_group()}}
\item \href{#method-ShintoMSALUser-set_group}{\code{shintoMSALUser$set_group()}}
\item \href{#method-ShintoMSALUser-remove_msal_account_from_app}{\code{shintoMSALUser$remove_msal_account_from_app()}}
\item \href{#method-ShintoMSALUser-has_role}{\code{shintoMSALUser$has_role()}}
\item \href{#method-ShintoMSALUser-is_admin}{\code{shintoMSALUser$is_admin()}}
\item \href{#method-ShintoMSALUser-get_application_roles}{\code{shintoMSALUser$get_application_roles()}}
\item \href{#method-ShintoMSALUser-get_application_groups}{\code{shintoMSALUser$get_application_groups()}}
\item \href{#method-ShintoMSALUser-log_timing}{\code{shintoMSALUser$log_timing()}}
\item \href{#method-ShintoMSALUser-add_invite}{\code{shintoMSALUser$add_invite()}}
\item \href{#method-ShintoMSALUser-has_invite}{\code{shintoMSALUser$has_invite()}}
\item \href{#method-ShintoMSALUser-get_invite}{\code{shintoMSALUser$get_invite()}}
\item \href{#method-ShintoMSALUser-update_invite}{\code{shintoMSALUser$update_invite()}}
\item \href{#method-ShintoMSALUser-remove_invite}{\code{shintoMSALUser$remove_invite()}}
\item \href{#method-ShintoMSALUser-make_msal_log}{\code{shintoMSALUser$make_msal_log()}}
\item \href{#method-ShintoMSALUser-get_applications}{\code{shintoMSALUser$get_applications()}}
\item \href{#method-ShintoMSALUser-clone}{\code{shintoMSALUser$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="append_data"><a href='../../shintodb/html/NA.html#method-NA-append_data'><code>shintodb::NA$append_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="close"><a href='../../shintodb/html/NA.html#method-NA-close'><code>shintodb::NA$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="connect_to_database"><a href='../../shintodb/html/NA.html#method-NA-connect_to_database'><code>shintodb::NA$connect_to_database()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="delete_rows_where"><a href='../../shintodb/html/NA.html#method-NA-delete_rows_where'><code>shintodb::NA$delete_rows_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="execute_query"><a href='../../shintodb/html/NA.html#method-NA-execute_query'><code>shintodb::NA$execute_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="filter"><a href='../../shintodb/html/NA.html#method-NA-filter'><code>shintodb::NA$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="get_query"><a href='../../shintodb/html/NA.html#method-NA-get_query'><code>shintodb::NA$get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_connection"><a href='../../shintodb/html/NA.html#method-NA-has_connection'><code>shintodb::NA$has_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_value"><a href='../../shintodb/html/NA.html#method-NA-has_value'><code>shintodb::NA$has_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="have_column"><a href='../../shintodb/html/NA.html#method-NA-have_column'><code>shintodb::NA$have_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_indexes"><a href='../../shintodb/html/NA.html#method-NA-list_indexes'><code>shintodb::NA$list_indexes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_tables"><a href='../../shintodb/html/NA.html#method-NA-list_tables'><code>shintodb::NA$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="make_column"><a href='../../shintodb/html/NA.html#method-NA-make_column'><code>shintodb::NA$make_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="postgres_now"><a href='../../shintodb/html/NA.html#method-NA-postgres_now'><code>shintodb::NA$postgres_now()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="postgres_now_in_local_time"><a href='../../shintodb/html/NA.html#method-NA-postgres_now_in_local_time'><code>shintodb::NA$postgres_now_in_local_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="query"><a href='../../shintodb/html/NA.html#method-NA-query'><code>shintodb::NA$query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="read_table"><a href='../../shintodb/html/NA.html#method-NA-read_table'><code>shintodb::NA$read_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where'><code>shintodb::NA$replace_value_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where2"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where2'><code>shintodb::NA$replace_value_where2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_columns"><a href='../../shintodb/html/NA.html#method-NA-table_columns'><code>shintodb::NA$table_columns()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_info"><a href='../../shintodb/html/NA.html#method-NA-table_info'><code>shintodb::NA$table_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="update_where"><a href='../../shintodb/html/NA.html#method-NA-update_where'><code>shintodb::NA$update_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="write_table_overwrite"><a href='../../shintodb/html/NA.html#method-NA-write_table_overwrite'><code>shintodb::NA$write_table_overwrite()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-new}{}}}
\subsection{Method \code{new()}}{
Make new shintousers object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$new(
  userid,
  appname = "",
  appversion = "",
  ad_groups = NULL,
  ad_authentication = FALSE,
  admin_group_pattern = "",
  con = NULL,
  pool = FALSE,
  dbusername = NULL,
  dbname = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{userid, if not NULL it is stored and used for all methods (handy inside an app)}

\item{\code{appname}}{application name}

\item{\code{appversion}}{Optional, application version string}

\item{\code{ad_groups}}{Optional, active directory-groups}

\item{\code{ad_authentication}}{Boolean that checks if active directory authentication has to be done}

\item{\code{admin_group_pattern}}{Optional pattern to see which groupnames are implicating admin users}

\item{\code{con}}{Optional, existing database connection to shintousers (for recycling)}

\item{\code{pool}}{Passed to [shintodb::connect()]}

\item{\code{dbusername}}{ignored ("shintousers")}

\item{\code{dbname}}{ignored ("shintousers")}

\item{\code{...}}{Further arguments passed to [shintodb::connect()]}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'shintousers' R6 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert to JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$to_json(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{An object to convert to JSON}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-from_json}{}}}
\subsection{Method \code{from_json()}}{
Convert to JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$from_json(txt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{txt}}{Text string, JSON}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-add_msal_user"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-add_msal_user}{}}}
\subsection{Method \code{add_msal_user()}}{
Add user to msal registry in shintousers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$add_msal_user(
  userid,
  appname,
  role = NA,
  groups = NA,
  username = NA,
  email,
  attributes = NA,
  comments = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{userid from MSAL for user}

\item{\code{appname}}{appname for the application user has access to}

\item{\code{role}}{Roles user has within application}

\item{\code{groups}}{JSON with groups user belongs to within application}

\item{\code{username}}{username (or label) for user}

\item{\code{email}}{email from user}

\item{\code{attributes}}{attributes from user}

\item{\code{comments}}{comments about user}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_name}{}}}
\subsection{Method \code{get_name()}}{
Read username for a user (default = current user)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_name(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{Vector of user ID's}

\item{\code{appname}}{Application name (can be NULL, appname on init is then used)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_user_name"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_user_name}{}}}
\subsection{Method \code{set_user_name()}}{
Sets the username for a user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_user_name(userid, appname, username)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{shiny application name}

\item{\code{username}}{The username, typically Voornaam Achternaam}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_email"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_email}{}}}
\subsection{Method \code{get_email()}}{
Read email address for a user (default = current user)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_email(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{Vector of user ID's}

\item{\code{appname}}{Application name (can be NULL, appname on init is then used)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_user_email"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_user_email}{}}}
\subsection{Method \code{set_user_email()}}{
Sets the email for a user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_user_email(userid, appname, email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{shiny application name}

\item{\code{email}}{The email}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_comment"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_comment}{}}}
\subsection{Method \code{get_comment()}}{
Read comment for a user (default = current user)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_comment(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{Vector of user ID's}

\item{\code{appname}}{Application name (can be NULL, appname on init is then used)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_comment"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_comment}{}}}
\subsection{Method \code{set_comment()}}{
Sets the comment for a user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_comment(userid, appname, comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}

\item{\code{comments}}{The comments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_last_login"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_last_login}{}}}
\subsection{Method \code{get_last_login()}}{
Get last login for this user for this application (reads `shintousers.logins`)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_last_login(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_last_login"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_last_login}{}}}
\subsection{Method \code{set_last_login()}}{
Update the last login for this user / appname (in table `shintousers.logins`)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_last_login(
  now = as.character(Sys.time()),
  userid,
  appname,
  appversion = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{now}}{Optional, time string (defaults to sys time)}

\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}

\item{\code{appversion}}{application version string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_and_set_last_login"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_and_set_last_login}{}}}
\subsection{Method \code{get_and_set_last_login()}}{
Reads last login, and sets the current time as the new 'last login'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_and_set_last_login(userid, appname, appversion = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}

\item{\code{appversion}}{application version string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If user has never logged in, writes a new line in `shintousers.logins`,
otherwise updates the last login.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-app_has_user"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-app_has_user}{}}}
\subsection{Method \code{app_has_user()}}{
Does an app have a user configured?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$app_has_user(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-list_application_users"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-list_application_users}{}}}
\subsection{Method \code{list_application_users()}}{
List users for an app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$list_application_users(
  appname,
  roles = NULL,
  groups = NULL,
  active_only = TRUE,
  ignore_groups = NULL,
  by_group = FALSE,
  add_last_login = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appname}}{application name}

\item{\code{roles}}{rolnames to filter on}

\item{\code{groups}}{groupnames to filter on}

\item{\code{active_only}}{Boolean to show if only active accounts need to be shown}

\item{\code{ignore_groups}}{Specific groups to ignore in the result. For example, exclude developer accounts}

\item{\code{by_group}}{Boolean, show by group in result}

\item{\code{add_last_login}}{Boolean, also show when user has logged in last.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_user_attributes"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_user_attributes}{}}}
\subsection{Method \code{set_user_attributes()}}{
Set attributes for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_user_attributes(userid, appname, attributes = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}

\item{\code{attributes}}{a list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_user_attributes"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_user_attributes}{}}}
\subsection{Method \code{get_user_attributes()}}{
Get attributes for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_user_attributes(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_role"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_role}{}}}
\subsection{Method \code{get_role()}}{
Gets the role for the current user (admin or viewer, typically)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_role(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_role"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_role}{}}}
\subsection{Method \code{set_role()}}{
Sets the role for a user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_role(userid, appname, role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}

\item{\code{role}}{The user role, typically 'admin', 'viewer'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_user_active_inactive"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_user_active_inactive}{}}}
\subsection{Method \code{get_user_active_inactive()}}{
Gets the activity status of a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_user_active_inactive(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_user_active_inactive"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_user_active_inactive}{}}}
\subsection{Method \code{set_user_active_inactive()}}{
Make a user active or inactive (sets 'active' field in 'roles' table)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_user_active_inactive(
  userid,
  appname,
  what = c("active", "inactive")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{User ID}

\item{\code{appname}}{Application name}

\item{\code{what}}{Either 'active' or 'inactive'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-disable_user"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-disable_user}{}}}
\subsection{Method \code{disable_user()}}{
Make the user inactive for an app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$disable_user(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{User ID}

\item{\code{appname}}{Application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-enable_user"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-enable_user}{}}}
\subsection{Method \code{enable_user()}}{
Make the user active for an app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$enable_user(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{User ID}

\item{\code{appname}}{Application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_group"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_group}{}}}
\subsection{Method \code{get_group()}}{
Get groups that the current user belongs to. See also `is_in_group`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_group(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-is_in_group"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-is_in_group}{}}}
\subsection{Method \code{is_in_group()}}{
Is the user in this group? `$is_in_group("superuser")` -> bool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$is_in_group(group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{Group name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-set_group"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-set_group}{}}}
\subsection{Method \code{set_group()}}{
Set the group for this user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$set_group(userid, appname, group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}

\item{\code{group}}{Group name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-remove_msal_account_from_app"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-remove_msal_account_from_app}{}}}
\subsection{Method \code{remove_msal_account_from_app()}}{
Remove a role for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$remove_msal_account_from_app(userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-has_role"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-has_role}{}}}
\subsection{Method \code{has_role()}}{
Does the current user have a role? `$has_role("admin")` --> bool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$has_role(role, userid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{The user role, typically 'admin', 'viewer'}

\item{\code{userid}}{msal userid}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-is_admin"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-is_admin}{}}}
\subsection{Method \code{is_admin()}}{
Better alternative to 'has_role("admin")' for use with AD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$is_admin()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_application_roles"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_application_roles}{}}}
\subsection{Method \code{get_application_roles()}}{
Get available roles for an application
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_application_roles(appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appname}}{The application name}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
!! Do not use in shiny applications (except shintousers_app) !!
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_application_groups"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_application_groups}{}}}
\subsection{Method \code{get_application_groups()}}{
Add an application to the list of applications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_application_groups(appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appname}}{The application name}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
!! Do not use in shiny applications (except shintousers_app) !!
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-log_timing"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-log_timing}{}}}
\subsection{Method \code{log_timing()}}{
Log a timing. Writes data to 'timings' table in shintousers, with appname, key and
double precision value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$log_timing(key, timing)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key in database table}

\item{\code{timing}}{Time in seconds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-add_invite"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-add_invite}{}}}
\subsection{Method \code{add_invite()}}{
Add an invitation for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$add_invite(
  email,
  username = NULL,
  invite_sent_by,
  appname,
  role = NA,
  groups = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{email of the user to be invited}

\item{\code{username}}{name of the user}

\item{\code{invite_sent_by}}{name of user that sent invite}

\item{\code{appname}}{appname of the application the invite is sent for}

\item{\code{role}}{the roles the invitee should get upon accepting}

\item{\code{groups}}{the groups the invitee should belong to upon accepting}

\item{\code{expire_date}}{The date on which the invite expires}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-has_invite"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-has_invite}{}}}
\subsection{Method \code{has_invite()}}{
Checks if there is an invite connected to email address for appname
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$has_invite(appname, email, ignore_expiration_date = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appname}}{application name}

\item{\code{email}}{email address of user to be checked}

\item{\code{ignore_expiration_date}}{Boolean that tells if a filter should be applied based on expiration date. Handy to find if any invite exits or a to find invites that are still valid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_invite"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_invite}{}}}
\subsection{Method \code{get_invite()}}{
Checks if there is an invite connected to email adress for appname
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_invite(appname, email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appname}}{application name}

\item{\code{email}}{email address of user to be checked}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-update_invite"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-update_invite}{}}}
\subsection{Method \code{update_invite()}}{
Update an invitation for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$update_invite(
  invite_id,
  email,
  username = NULL,
  invite_sent_by,
  appname,
  role = NA,
  groups = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{invite_id}}{The id of the invitation in de the table shiny_msal_accounts_pending_invites}

\item{\code{email}}{email of the user to be invited}

\item{\code{username}}{name of the user}

\item{\code{invite_sent_by}}{name of user that sent invite}

\item{\code{appname}}{appname of the application the invite is sent for}

\item{\code{role}}{the roles the invitee should get upon accepting}

\item{\code{groups}}{the groups the invitee should belong to upon accepting}

\item{\code{expire_date}}{The date on which the invite expires}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-remove_invite"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-remove_invite}{}}}
\subsection{Method \code{remove_invite()}}{
Remove an invitation for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$remove_invite(inviteid, appname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inviteid}}{id of invitation to be deleted}

\item{\code{appname}}{application name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-make_msal_log"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-make_msal_log}{}}}
\subsection{Method \code{make_msal_log()}}{
Add a logging entry for MSAL users
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$make_msal_log(user_id, appname, action_type, logging_message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{id of user causing the action that is logged}

\item{\code{appname}}{application name}

\item{\code{action_type}}{Categorical type of action}

\item{\code{logging_message}}{Detailed logging message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-get_applications"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-get_applications}{}}}
\subsection{Method \code{get_applications()}}{
Get list of available applications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$get_applications()}\if{html}{\out{</div>}}
}

\subsection{Details}{
!! Do not use in shiny applications (except shintousers_app) !!
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShintoMSALUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShintoMSALUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoMSALUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
