% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_r6_object.R
\name{shintoUser}
\alias{shintoUser}
\title{A User object for use in Shinto apps}
\description{
A User object for use in Shinto apps

A User object for use in Shinto apps
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{shintoUser$new()}}
\item \href{#method-query}{\code{shintoUser$query()}}
\item \href{#method-execute_query}{\code{shintoUser$execute_query()}}
\item \href{#method-append_data}{\code{shintoUser$append_data()}}
\item \href{#method-read_table}{\code{shintoUser$read_table()}}
\item \href{#method-get_last_login}{\code{shintoUser$get_last_login()}}
\item \href{#method-set_last_login}{\code{shintoUser$set_last_login()}}
\item \href{#method-get_and_set_last_login}{\code{shintoUser$get_and_set_last_login()}}
\item \href{#method-get_role}{\code{shintoUser$get_role()}}
\item \href{#method-set_role}{\code{shintoUser$set_role()}}
\item \href{#method-remove_role}{\code{shintoUser$remove_role()}}
\item \href{#method-has_role}{\code{shintoUser$has_role()}}
\item \href{#method-add_application}{\code{shintoUser$add_application()}}
\item \href{#method-set_application_roles}{\code{shintoUser$set_application_roles()}}
\item \href{#method-get_application_roles}{\code{shintoUser$get_application_roles()}}
\item \href{#method-get_applications}{\code{shintoUser$get_applications()}}
\item \href{#method-remove_application}{\code{shintoUser$remove_application()}}
\item \href{#method-clone}{\code{shintoUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$new(dbusername, userid = NULL, appname = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$query(txt, glue = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute_query"></a>}}
\if{latex}{\out{\hypertarget{method-execute_query}{}}}
\subsection{Method \code{execute_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$execute_query(txt, glue = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_data"></a>}}
\if{latex}{\out{\hypertarget{method-append_data}{}}}
\subsection{Method \code{append_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$append_data(table, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_table"></a>}}
\if{latex}{\out{\hypertarget{method-read_table}{}}}
\subsection{Method \code{read_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$read_table(table, lazy = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_last_login"></a>}}
\if{latex}{\out{\hypertarget{method-get_last_login}{}}}
\subsection{Method \code{get_last_login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$get_last_login(userid = NULL, appname = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_last_login"></a>}}
\if{latex}{\out{\hypertarget{method-set_last_login}{}}}
\subsection{Method \code{set_last_login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$set_last_login(now = as.character(Sys.time()))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_and_set_last_login"></a>}}
\if{latex}{\out{\hypertarget{method-get_and_set_last_login}{}}}
\subsection{Method \code{get_and_set_last_login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$get_and_set_last_login()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_role"></a>}}
\if{latex}{\out{\hypertarget{method-get_role}{}}}
\subsection{Method \code{get_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$get_role(userid = NULL, appname = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_role"></a>}}
\if{latex}{\out{\hypertarget{method-set_role}{}}}
\subsection{Method \code{set_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$set_role(userid, appname, role)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_role"></a>}}
\if{latex}{\out{\hypertarget{method-remove_role}{}}}
\subsection{Method \code{remove_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$remove_role(userid, appname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_role"></a>}}
\if{latex}{\out{\hypertarget{method-has_role}{}}}
\subsection{Method \code{has_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$has_role(role)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_application"></a>}}
\if{latex}{\out{\hypertarget{method-add_application}{}}}
\subsection{Method \code{add_application()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$add_application(appname, roles, comment = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_application_roles"></a>}}
\if{latex}{\out{\hypertarget{method-set_application_roles}{}}}
\subsection{Method \code{set_application_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$set_application_roles(appname, roles)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_application_roles"></a>}}
\if{latex}{\out{\hypertarget{method-get_application_roles}{}}}
\subsection{Method \code{get_application_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$get_application_roles(appname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_applications"></a>}}
\if{latex}{\out{\hypertarget{method-get_applications}{}}}
\subsection{Method \code{get_applications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$get_applications()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_application"></a>}}
\if{latex}{\out{\hypertarget{method-remove_application}{}}}
\subsection{Method \code{remove_application()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$remove_application(appname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{shintoUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
