% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_db_connection.R
\name{users_db_connection}
\alias{users_db_connection}
\title{Connect to the users database from an app}
\usage{
users_db_connection(dbusername, config_file = "conf/config.yml")
}
\arguments{
\item{dbusername}{Database user name on devpostgres02 for this app ('risicoradar','wbm_eindhoven')}
}
\description{
Connect to the users database from an app
}
